♻️ Move buildLabelSelector to shared package to eliminate duplication
