+++ shared/src/index.ts
@@ -1,3 +1,4 @@
 export * from "./types/index";
 export * from "./transformation";
 export * from "./effort";
+export * from "./labelSelector";
