+++ webview-ui/src/components/ProfileManager/ProfileEditorForm.tsx
@@ -31,7 +31,7 @@ import { useExtensionStateContext } from "../../context/ExtensionStateContext";
 import { AnalysisProfile, CONFIGURE_CUSTOM_RULES } from "@editor-extensions/shared";
 import { ConfirmDialog } from "../ConfirmDialog/ConfirmDialog";
 import { CreatableMultiSelectField } from "./CreatableMultiSelectField";
-import { buildLabelSelector } from "../../utils/labelSelector";
+import { buildLabelSelector } from "@editor-extensions/shared";
 
 function useDebouncedCallback(callback: (...args: any[]) => void, delay: number) {
   const timeoutRef = useRef<NodeJS.Timeout | null>(null);
